@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -31,26 +31,26 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.5
-  freezed_annotation: ^2.2.0
+  cupertino_icons: ^1.0.6
+  freezed_annotation: ^2.4.1
   json_annotation: ^4.8.1
   transparent_image: ^2.0.1
-  dio: ^5.2.1+1
-  flutter_riverpod: ^2.3.6
-  riverpod_annotation: ^2.1.1
-  envied: ^0.3.0+3
+  dio: ^5.3.4
+  flutter_riverpod: ^2.4.9
+  riverpod_annotation: ^2.3.3
+  envied: ^0.5.2
   shimmer: ^3.0.0
-  cached_network_image: ^3.2.3
-  go_router: ^9.0.0
+  cached_network_image: ^3.3.0
+  go_router: ^12.1.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.4.5
-  riverpod_generator: ^2.2.3
-  freezed: ^2.3.5
-  json_serializable: ^6.7.0
-  envied_generator: ^0.3.0+3
+  build_runner: ^2.4.6
+  riverpod_generator: ^2.3.9
+  freezed: ^2.4.5
+  json_serializable: ^6.7.1
+  envied_generator: ^0.5.2
 
   
   # The "flutter_lints" package below contains a set of recommended lints to
@@ -58,7 +58,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.2
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
