@@ -3,7 +3,7 @@ import 'package:envied/envied.dart';
 part 'env.g.dart';
 
 @Envied(path: '.env')
-abstract class Env {
+final class Env {
   @EnviedField(varName: 'TMDB_KEY', obfuscate: true)
-  static final tmdbApiKey = _Env.tmdbApiKey;
+  static final String tmdbApiKey = _Env.tmdbApiKey;
 }
